+++ __tests__/server/utils/loadModules.spec.jsx
@@ -20,6 +20,7 @@ import { CONFIGURATION_KEY } from '../../../src/server/utils/onModuleLoad';
 import loadModules from '../../../src/server/utils/loadModules';
 import { updateCSP } from '../../../src/server/middleware/csp';
 import { setClientModuleMapCache, getClientModuleMapCache } from '../../../src/server/utils/clientModuleMapCache';
+import addBaseUrlToModuleMap from '../../../src/server/utils/addBaseUrlToModuleMap';
 
 // This named export exists only on the mock
 // eslint-disable-next-line import/named
@@ -95,7 +96,7 @@ describe('loadModules', () => {
   it('updates the holocron module registry', async () => {
     await loadModules();
     expect(updateModuleRegistry).toHaveBeenCalledWith({
-      moduleMap,
+      moduleMap: addBaseUrlToModuleMap(moduleMap),
       batchModulesToUpdate: require('../../../src/server/utils/batchModulesToUpdate').default,
       getModulesToUpdate: require('../../../src/server/utils/getModulesToUpdate').default,
       onModuleLoad: require('../../../src/server/utils/onModuleLoad').default,