+++ src/client/service-worker/events/index.js
@@ -15,3 +15,4 @@
  */
 
 export { createInstallMiddleware, createActivateMiddleware } from './lifecycle';
+export { default as createFetchMiddleware } from './fetch';