+++ src/server/ssrServer.js
@@ -44,7 +44,11 @@ import checkStateForStatusCode from './middleware/checkStateForStatusCode';
 import sendHtml, { renderStaticErrorPage } from './middleware/sendHtml';
 import logging from './utils/logging/serverMiddleware';
 import forwardedHeaderParser from './middleware/forwardedHeaderParser';
-import { serviceWorkerMiddleware, webManifestMiddleware } from './middleware/pwa';
+import {
+  serviceWorkerMiddleware,
+  webManifestMiddleware,
+  offlineMiddleware,
+} from './middleware/pwa';
 
 export function createApp({ enablePostToModuleRoutes = false } = {}) {
   const app = express();
@@ -73,6 +77,7 @@ export function createApp({ enablePostToModuleRoutes = false } = {}) {
   app.post('/_/report/errors', clientErrorLogger);
   app.get('**/*.(json|js|css|map)', (req, res) => res.sendStatus(404));
 
+  app.get('/_/pwa/shell', offlineMiddleware(oneApp));
   app.get(
     '*',
     addFrameOptionsHeader,