+++ src/server/utils/clientModuleMapCache.js
@@ -16,6 +16,8 @@
 
 let cache = {};
 
+/* Filters bundle types so only the required client moduleBundleType is returned
+in ../server/middleware/sendHtml.js and reducing the html payload size */
 function filterBundles(moduleMap, moduleBundleType) {
   return {
     ...moduleMap,